first major attempt at refactoring
