+++ application/static/style.css
@@ -1,0 +1,211 @@
+  .row {
+    background: #ffffff;
+    margin-left: 0px;
+    margin-top: 0px;
+    text-align: left;
+  }
+  
+  .col {
+    border: solid 0px #6c757d; /* set this to 1 to see the layout */
+    padding: 0px;
+  }
+  
+  .description {
+    padding: 5px;
+  }
+
+  .figure1 {
+    float: right;
+  }
+  
+  #summary {
+    max-height: 100px;
+    overflow: hidden;
+  }
+
+  .search-title {
+    height: 3.3em;
+    line-height: 1.5;
+    font-size: 1.3em;
+    margin-top: 0;
+    margin-bottom: 0;
+    margin-left: 0;
+    margin-right: 0;
+    font-weight: bold;
+    overflow: hidden;
+    /* white-space: nowrap;
+    text-overflow: ellipsis; */
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;  
+  }
+  
+  .ellipsis {
+    text-overflow: ellipsis;
+  
+    /* Required for text-overflow to do anything */
+    white-space: wrap;
+  }
+
+  .sort_by {
+    display: flex;
+    align-items: flex-end;
+    justify-content: flex-end;
+  }
+
+  .container-flex {
+    display: flex;
+  }
+
+  .container-search-result {
+    margin-top: 24px;
+    margin-right: 0px;
+  }
+
+  .container-main {
+    max-width: 1200px;
+  }
+
+  .left-padded {
+    padding-left: 60px;
+  }
+
+  #emissions-histogram {
+    max-height: 100%;
+    max-width: 100%;
+  }
+
+  .ingredient-list {
+    list-style-type: circle;
+    list-style-position: outside;
+  }
+
+  .directions-list {
+    list-style-type: decimal;
+    line-height: 110%;
+    list-style-position: outside;
+  }
+
+  .nutrition {
+    text-align: left;
+  }
+
+  .servings {
+    text-align: left;
+  }
+
+  .next-page {
+    text-align: center;
+  }
+
+  .icon {
+    border: 20px;
+  }
+  
+  /* star rating */
+  .stars-outer {
+    position: relative;
+    display: inline-block;
+  }
+
+  .sustainability-outer {
+    position: relative;
+    display: inline-block;
+  }
+
+  .similarity-outer {
+    position: relative;
+    display: inline-block;
+  }
+  
+  .stars-inner {
+    position: absolute;
+    top: 0;
+    left: 0;
+    white-space: nowrap;
+    overflow: hidden;
+    width: 0;
+  }
+
+  .sustainability-inner {
+    position: absolute;
+    top: 0;
+    left: 0;
+    white-space: nowrap;
+    overflow: hidden;
+    width: 0;
+  }
+
+  .similarity-inner {
+    position: absolute;
+    top: 0;
+    left: 0;
+    white-space: nowrap;
+    overflow: hidden;
+    width: 0;
+  }
+  
+  .stars-outer::before {
+    content: "\f005 \f005 \f005 \f005 \f005";
+    font-family: "Font Awesome 5 Free";
+    font-weight: 900;
+    color: #ccc;
+  }
+  
+  .stars-inner::before {
+    content: "\f005 \f005 \f005 \f005 \f005";
+    font-family: "Font Awesome 5 Free";
+    font-weight: 900;
+    color: #f8ce0b;
+  }
+  
+  .sustainability-outer::before {
+    content: "\f06c \f06c \f06c \f06c \f06c"; /* "\f0c2 \f0c2 \f0c2 \f0c2 \f0c2"; */
+    font-family: "Font Awesome 5 Free";
+    font-weight: 900;
+    color: #ccc;
+  }
+
+  .sustainability-inner::before {
+    content: "\f06c \f06c \f06c \f06c \f06c"; /* "\f0c2 \f0c2 \f0c2 \f0c2 \f0c2"; */
+    font-family: "Font Awesome 5 Free";
+    font-weight: 900;
+    color: #12d800;
+  }
+
+  .similarity-outer::before {
+    content: "\f291 \f291 \f291 \f291 \f291";
+    font-family: "Font Awesome 5 Free";
+    font-weight: 900;
+    color: #ccc;
+  }
+
+  .similarity-inner::before {
+    content: "\f291 \f291 \f291 \f291 \f291";
+    font-family: "Font Awesome 5 Free";
+    font-weight: 900;
+    color: #1851f0;
+  }
+  
+#form-group {
+    position: fixed;
+    top: 0px;
+    left: 0;
+    z-index: 999;
+    width: 100%;
+    height: 100vh;
+    text-align: center;
+    background-color: rgba(0,0,0,0.3);
+}
+
+.blog_iframe {
+  overflow: hidden;
+  height: 86vh;
+  width: 100%;
+  display:block;
+}
+
+.logo_startpage {
+  width: 75px;
+  height: 75px;
+}
