+++ application/templates/field-macro.html
@@ -1,0 +1,14 @@
+{# Render form fields #}
+{% macro displayField(field, fieldName) %}
+    <div class="form-group">
+        <div class="input-group mb-3">
+            {{ field(class_='form-control', placeholder=fieldName, **kwargs) }}
+            <!-- Error messages -->
+            <ul class="formError">
+                {% for error in field.errors %}
+                    <li>"{{ error }}"</li>
+                {% endfor %}
+            </ul>
+        </div>
+    </div>
+{% endmacro %}
