+++ application/templates/signup.html
@@ -1,0 +1,26 @@
+{% from 'field-macro.html' import displayField %}
+{% extends 'base.html' %}
+{% block title %} Recipes searcher {% endblock %}
+{% block body %}
+<div class="row justify-content-center">
+    <div class="col-12 col-xl-6">
+        <center>
+            <br><br>
+            <h2>Create a new account</h2><br>
+            <div id="container">
+                <form action="{{ url_for('signup') }}", method="POST">
+                    {{ displayField(reg_form.username, 'Username', autocomplete="off", autofocus=true) }}
+                    {{ displayField(reg_form.password, 'Password') }}
+                    {{ displayField(reg_form.confirm_pswd, 'Retype Password') }}
+                    <div class="form-group">
+                        <input type="submit" value="Create" class="btn btn-warning">
+                    </div>
+                    {{ reg_form.csrf_token }}
+                </form> 
+            </div>
+            <hr>
+            <p>Already registered? <a href="{{ url_for('signin') }}">Login here</a>.</p>
+        </center>
+    </div>
+</div>
+{% endblock %}
